# Save compression: 10-15x+ smaller maps, faster saves, corruption resistant, also makes julienne fries.