--- a/msvc-full-features/zstd.vcxproj
+++ b/msvc-full-features/zstd.vcxproj
@@ -1,0 +1,177 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Project DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <ItemGroup Label="ProjectConfigurations">
+    <ProjectConfiguration Include="Debug|ARM64EC">
+      <Configuration>Debug</Configuration>
+      <Platform>ARM64EC</Platform>
+    </ProjectConfiguration>
+    <ProjectConfiguration Include="Debug|Win32">
+      <Configuration>Debug</Configuration>
+      <Platform>Win32</Platform>
+    </ProjectConfiguration>
+    <ProjectConfiguration Include="Release|ARM64EC">
+      <Configuration>Release</Configuration>
+      <Platform>ARM64EC</Platform>
+    </ProjectConfiguration>
+    <ProjectConfiguration Include="Release|Win32">
+      <Configuration>Release</Configuration>
+      <Platform>Win32</Platform>
+    </ProjectConfiguration>
+    <ProjectConfiguration Include="Debug|x64">
+      <Configuration>Debug</Configuration>
+      <Platform>x64</Platform>
+    </ProjectConfiguration>
+    <ProjectConfiguration Include="Release|x64">
+      <Configuration>Release</Configuration>
+      <Platform>x64</Platform>
+    </ProjectConfiguration>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="..\src\third-party\zstd\common\allocations.h" />
+    <ClInclude Include="..\src\third-party\zstd\common\bits.h" />
+    <ClInclude Include="..\src\third-party\zstd\common\bitstream.h" />
+    <ClInclude Include="..\src\third-party\zstd\common\compiler.h" />
+    <ClInclude Include="..\src\third-party\zstd\common\cpu.h" />
+    <ClInclude Include="..\src\third-party\zstd\common\debug.h" />
+    <ClInclude Include="..\src\third-party\zstd\common\error_private.h" />
+    <ClInclude Include="..\src\third-party\zstd\common\fse.h" />
+    <ClInclude Include="..\src\third-party\zstd\common\huf.h" />
+    <ClInclude Include="..\src\third-party\zstd\common\mem.h" />
+    <ClInclude Include="..\src\third-party\zstd\common\pool.h" />
+    <ClInclude Include="..\src\third-party\zstd\common\portability_macros.h" />
+    <ClInclude Include="..\src\third-party\zstd\common\threading.h" />
+    <ClInclude Include="..\src\third-party\zstd\common\xxhash.h" />
+    <ClInclude Include="..\src\third-party\zstd\common\zstd_deps.h" />
+    <ClInclude Include="..\src\third-party\zstd\common\zstd_internal.h" />
+    <ClInclude Include="..\src\third-party\zstd\common\zstd_trace.h" />
+    <ClInclude Include="..\src\third-party\zstd\compress\clevels.h" />
+    <ClInclude Include="..\src\third-party\zstd\compress\hist.h" />
+    <ClInclude Include="..\src\third-party\zstd\compress\zstdmt_compress.h" />
+    <ClInclude Include="..\src\third-party\zstd\compress\zstd_compress_internal.h" />
+    <ClInclude Include="..\src\third-party\zstd\compress\zstd_compress_literals.h" />
+    <ClInclude Include="..\src\third-party\zstd\compress\zstd_compress_sequences.h" />
+    <ClInclude Include="..\src\third-party\zstd\compress\zstd_compress_superblock.h" />
+    <ClInclude Include="..\src\third-party\zstd\compress\zstd_cwksp.h" />
+    <ClInclude Include="..\src\third-party\zstd\compress\zstd_double_fast.h" />
+    <ClInclude Include="..\src\third-party\zstd\compress\zstd_fast.h" />
+    <ClInclude Include="..\src\third-party\zstd\compress\zstd_lazy.h" />
+    <ClInclude Include="..\src\third-party\zstd\compress\zstd_ldm.h" />
+    <ClInclude Include="..\src\third-party\zstd\compress\zstd_ldm_geartab.h" />
+    <ClInclude Include="..\src\third-party\zstd\compress\zstd_opt.h" />
+    <ClInclude Include="..\src\third-party\zstd\compress\zstd_preSplit.h" />
+    <ClInclude Include="..\src\third-party\zstd\decompress\zstd_ddict.h" />
+    <ClInclude Include="..\src\third-party\zstd\decompress\zstd_decompress_block.h" />
+    <ClInclude Include="..\src\third-party\zstd\decompress\zstd_decompress_internal.h" />
+    <ClInclude Include="..\src\third-party\zstd\zstd.h" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClCompile Include="..\src\third-party\zstd\common\debug.c" />
+    <ClCompile Include="..\src\third-party\zstd\common\entropy_common.c" />
+    <ClCompile Include="..\src\third-party\zstd\common\error_private.c" />
+    <ClCompile Include="..\src\third-party\zstd\common\fse_decompress.c" />
+    <ClCompile Include="..\src\third-party\zstd\common\pool.c" />
+    <ClCompile Include="..\src\third-party\zstd\common\threading.c" />
+    <ClCompile Include="..\src\third-party\zstd\common\xxhash.c" />
+    <ClCompile Include="..\src\third-party\zstd\common\zstd_common.c" />
+    <ClCompile Include="..\src\third-party\zstd\compress\fse_compress.c" />
+    <ClCompile Include="..\src\third-party\zstd\compress\hist.c" />
+    <ClCompile Include="..\src\third-party\zstd\compress\huf_compress.c" />
+    <ClCompile Include="..\src\third-party\zstd\compress\zstdmt_compress.c" />
+    <ClCompile Include="..\src\third-party\zstd\compress\zstd_compress.c" />
+    <ClCompile Include="..\src\third-party\zstd\compress\zstd_compress_literals.c" />
+    <ClCompile Include="..\src\third-party\zstd\compress\zstd_compress_sequences.c" />
+    <ClCompile Include="..\src\third-party\zstd\compress\zstd_compress_superblock.c" />
+    <ClCompile Include="..\src\third-party\zstd\compress\zstd_double_fast.c" />
+    <ClCompile Include="..\src\third-party\zstd\compress\zstd_fast.c" />
+    <ClCompile Include="..\src\third-party\zstd\compress\zstd_lazy.c" />
+    <ClCompile Include="..\src\third-party\zstd\compress\zstd_ldm.c" />
+    <ClCompile Include="..\src\third-party\zstd\compress\zstd_opt.c" />
+    <ClCompile Include="..\src\third-party\zstd\compress\zstd_preSplit.c" />
+    <ClCompile Include="..\src\third-party\zstd\decompress\huf_decompress.c" />
+    <ClCompile Include="..\src\third-party\zstd\decompress\zstd_ddict.c" />
+    <ClCompile Include="..\src\third-party\zstd\decompress\zstd_decompress.c" />
+    <ClCompile Include="..\src\third-party\zstd\decompress\zstd_decompress_block.c" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="..\src\third-party\zstd\decompress\huf_decompress_amd64.S" />
+  </ItemGroup>
+  <PropertyGroup Label="Globals">
+    <VCProjectVersion>16.0</VCProjectVersion>
+    <Keyword>Win32Proj</Keyword>
+    <ProjectGuid>{f49f14c6-b046-41c7-b707-03366f4b1578}</ProjectGuid>
+    <RootNamespace>flatbuffers</RootNamespace>
+    <WindowsTargetPlatformVersion>10.0</WindowsTargetPlatformVersion>
+  </PropertyGroup>
+  <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
+  <PropertyGroup Label="Configuration">
+    <ConfigurationType>StaticLibrary</ConfigurationType>
+    <PlatformToolset>$(DefaultPlatformToolset)</PlatformToolset>
+    <CharacterSet>Unicode</CharacterSet>
+    <PreferredToolArchitecture>x64</PreferredToolArchitecture>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(Configuration)'=='Debug'" Label="Configuration">
+    <UseDebugLibraries>true</UseDebugLibraries>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(Configuration)'=='Release'" Label="Configuration">
+    <UseDebugLibraries>false</UseDebugLibraries>
+    <WholeProgramOptimization>false</WholeProgramOptimization>
+  </PropertyGroup>
+  <Import Project="Cataclysm.Cpp.props" />
+  <Import Project="$(VCTargetsPath)\Microsoft.Cpp.props" />
+  <ImportGroup Label="ExtensionSettings">
+  </ImportGroup>
+  <ImportGroup Label="Shared">
+  </ImportGroup>
+  <ImportGroup Label="PropertySheets">
+    <Import Project="$(UserRootDir)\Microsoft.Cpp.$(Platform).user.props" Condition="exists('$(UserRootDir)\Microsoft.Cpp.$(Platform).user.props')" Label="LocalAppDataPlatform" />
+    <Import Project="Cataclysm-common.props" />
+  </ImportGroup>
+  <PropertyGroup Label="UserMacros" />
+  <PropertyGroup Condition="'$(Configuration)'=='Debug'">
+    <LinkIncremental>true</LinkIncremental>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(Configuration)'=='Release'">
+    <LinkIncremental>false</LinkIncremental>
+  </PropertyGroup>
+  <PropertyGroup Label="Vcpkg">
+    <VcpkgEnabled>false</VcpkgEnabled>
+  </PropertyGroup>
+  <ItemDefinitionGroup>
+    <ClCompile>
+      <PreprocessorDefinitions>_CONSOLE;ZSTD_DISABLE_ASM;%(PreprocessorDefinitions)</PreprocessorDefinitions>
+    </ClCompile>
+  </ItemDefinitionGroup>
+  <ItemDefinitionGroup Condition="'$(Configuration)'=='Debug'">
+    <ClCompile>
+      <WarningLevel>Level3</WarningLevel>
+      <FunctionLevelLinking>true</FunctionLevelLinking>
+      <PreprocessorDefinitions>_DEBUG;%(PreprocessorDefinitions)</PreprocessorDefinitions>
+      <RuntimeLibrary>MultiThreadedDebug</RuntimeLibrary>
+    </ClCompile>
+    <Link>
+      <EnableCOMDATFolding>true</EnableCOMDATFolding>
+      <OptimizeReferences>true</OptimizeReferences>
+    </Link>
+  </ItemDefinitionGroup>
+  <ItemDefinitionGroup Condition="'$(Configuration)'=='Release'">
+    <ClCompile>
+      <WarningLevel>Level3</WarningLevel>
+      <FunctionLevelLinking>true</FunctionLevelLinking>
+      <IntrinsicFunctions>true</IntrinsicFunctions>
+      <PreprocessorDefinitions>NDEBUG;%(PreprocessorDefinitions)</PreprocessorDefinitions>
+      <RuntimeLibrary>MultiThreaded</RuntimeLibrary>
+    </ClCompile>
+    <Link>
+      <EnableCOMDATFolding>true</EnableCOMDATFolding>
+      <OptimizeReferences>true</OptimizeReferences>
+    </Link>
+  </ItemDefinitionGroup>
+  <ItemDefinitionGroup Condition="'$(Platform)'=='Win32'">
+    <ClCompile>
+      <PreprocessorDefinitions>WIN32;%(PreprocessorDefinitions)</PreprocessorDefinitions>
+    </ClCompile>
+  </ItemDefinitionGroup>
+  <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
+  <ImportGroup Label="ExtensionTargets">
+  </ImportGroup>
+</Project>