--- a/src/auto_pickup.cpp
+++ b/src/auto_pickup.cpp
@@ -889,8 +889,9 @@ bool player_settings::save( const bool bCharacter )
         savefile = PATH_INFO::player_base_save_path() + ".apu.json";
 
         const cata_path player_save = PATH_INFO::player_base_save_path() + ".sav";
+        const cata_path player_save_zzip = player_save + ".zzip";
         //Character not saved yet.
-        if( !file_exist( player_save ) ) {
+        if( !file_exist( player_save ) || !file_exist( player_save_zzip ) ) {
             return true;
         }
     }