--- a/src/cata_utility.cpp
+++ b/src/cata_utility.cpp
@@ -35,6 +35,7 @@
 #include "unicode.h"
 #include "zlib.h"
 #include "zzip.h"
+#include "zzip_stack.h"
 
 static double pow10( unsigned int n )
 {
@@ -605,6 +606,25 @@ bool read_from_file_optional_json( const cata_path &path,
 }
 
 bool read_from_zzip_optional( const std::shared_ptr<zzip> &z,
+                              const std::filesystem::path &file,
+                              const std::function<void( std::string_view )> &reader )
+{
+    if( !z || !z->has_file( file ) ) {
+        return false;
+    }
+    try {
+        std::vector<std::byte> file_data = z->get_file( file );
+        std::string_view sv{ reinterpret_cast<const char *>( file_data.data() ), file_data.size() };
+        reader( sv );
+        return true;
+    } catch( const std::exception &err ) {
+        debugmsg( _( "Failed to read from \"%1$s\": %2$s" ), file.generic_u8string().c_str(),
+                  err.what() );
+        return false;
+    }
+}
+
+bool read_from_zzip_optional( const std::shared_ptr<zzip_stack> &z,
                               const std::filesystem::path &file,
                               const std::function<void( std::string_view )> &reader )
 {