… android/app/jni/src/Android.mk
… android/app/jni/src/third-party/Android.mk
… data/raw/compression/maps.dict
… data/raw/compression/mmr.dict
… data/raw/compression/overmaps.dict
… msvc-full-features/Cataclysm-vcpkg-static.sln
… src/game.h
… src/global_vars.h
… src/json_loader.cpp
… src/json_loader.h
… src/main_menu.cpp
… src/map_memory.cpp
… src/mapbuffer.cpp
… src/mmap_file.cpp
… src/mmap_file.h
… src/options.cpp
… src/overmap.cpp
… src/overmap.h
… src/overmapbuffer.cpp
… src/path_info.cpp
… src/path_info.h
… src/perf.cpp
… src/perf.h
… src/safemode_ui.cpp
… src/savegame.cpp
… src/std_hash_fs_path.h
… src/third-party/CMakeLists.txt
… src/translation_document.cpp
… src/translation_document.h
… src/worldfactory.cpp