<!-- HOW TO USE: Under each "#### Heading" below, enter information relevant to your pull request.
Leave the headings unless they don't apply to your PR.

Please read carefully.
Once a pull request is submitted, automatic stylistic and consistency checks will be performed on the PR's changes.
The results can be either seen under the "Files changed" section of a PR or in the check's details.

Rules for suggested pull requests:
- If possible, limit yourself to small changes, 500 strings at max. Exceptions are adding or changing maps, and changes, that won't work unless they are done in a single run (even then there can be ways) - violating it puts a lot of unnecessary work on our merge team.
- Do not scope creep. If you make a pull request "Add new gun", please do not make anything more than adding the gun and following changes, like changing the stats of the gun, removing other guns from itemgroups or tweaking zombie horse stats - violating it makes future search and debugging stuff much harder, since PR name is not related to what is changed in the game. "Who the hell removed the quest item from drop in location X in PR, that adds a new plushie" - this may be a quote from a person who was affected by scope creep
- Do not make omnibus PRs. Meaning do not make a single PR, that fixes ten different, not related issues, at once, even if they are all one string - same as previously mentioned scope creep, it doesn't help to search the changes when debugging, despite all power of git blame tool

NOTE: Please grant permission for repository maintainers to edit your PR.  It is EXTREMELY common for PRs to be held up due to trivial changes being requested and the author being unavailable to make them. -->

#### Summary
Infrastructure "Zstandard compressed save data for disk space savings."
<!-- This section should consist of exactly one line, edit the one above.
1. Replace the word "Category" with one of these words: Features, Content, Interface, Mods, Balance, Bugfixes, Performance, Infrastructure, Build, I18N.
2. Replace the text inside the quotes with a brief description of your changes.
Or if you don't want a changelog entry, replace the whole line with just the word "None" (with no quotes).
Examples:
1. None
2. Features "In-game Armor sprite change"
3. Interface "Show crafting failure chances in the crafting interface"
4. Infrastructure "JSON-ize slot machines"
5. Bugfixes "Crafting GUI: show how much recipe makes for non-charge items"
For more on the meaning of each category, see:
https://github.com/CleverRaven/Cataclysm-DDA/blob/master/doc/CHANGELOG_GUIDELINES.md
If approved and merged, your summary will be added to the project changelog:
https://github.com/CleverRaven/Cataclysm-DDA/blob/master/data/changelog.txt -->

#### Purpose of change
Long games get big. Very big. I've been given a sample of one which is almost 4GB of just `maps/` data, not including map memory. Sky Islands mod type games also generate lots of terrain. Multiple dimensions are coming. All of these incur more disk pressure from save data. Compression is the natural answer for this, because save data is almost never meant to be read or modified by humans anyway.
<!-- With a few sentences, describe your reasons for making this change.
If it relates to an existing issue, you can link it with a # followed by the GitHub issue number, like #1234.
When you submit a pull request that completely resolves an issue, use [Github's closing keywords](https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/using-keywords-in-issues-and-pull-requests#linking-a-pull-request-to-an-issue)
to automatically close the issue once your pull request is merged.
If there is no related issue, explain here what issue, feature, or other concern you are addressing.  If this is a bugfix, include steps to reproduce the original bug, so your fix can be verified. -->

#### Describe the solution
The solution is multipart.

- First, pick a compression algorithm.
  - gzip is terrible and slow.
  - lzma is good but slow.
  - Zstandard (`zstd`) is a modern compression algorithm that generally beats both at speed or performance, and sometimes both at once. It seemed like a plausible option to shoot for first. The library source is not large and is straightforward to integrate with the project due to only having internal dependencies.
  - I chose `zstd`.

- Next, pick a format. Part of the problem with saves is the many large files which individually end up less than a page large, but because modern hard disks have a minimum allocation size of 4KB end up wasting large amounts of disk. So some sort of archive format like zip or tar should be used.
  - tar is not suitable because it does not support random access to compressed entries. Map data is aggregated in folders where only subsets are loaded / accessed at a time. Linearly decompressing from the beginning would be slow and wasteful.
  - zip supports random access, but I could not find a good open source library for it.
  - [dar](http://dar.linux.free.fr/) was considered as it does support random access to entries, but overall it seems like an extremely overkill library as it is intended for _disk_ archiving, not just a grabbag of files.
  - So I settled on a proprietary format. There are many advantages. They will be detailed in `Additional context`.

The logistics of managing compressed saves is relatively straightforward. `zstd` supports using dictionaries for best compression ratios of small data. The presence of these dictionaries inside the world folder is used as a proxy for whether compression is enabled or not. This also means that save portability is trivial - the dictionary is encoded per-save, and is independent of the dictionary in the repo itself. Users can substitute their own dictionaries without hassle if desired.

Implementing the format was the hard part. Once it's done, it's trivial to wire up in the appropriate save code paths.

<!-- How does the feature work, or how does this fix a bug?  The easier you make your solution to understand, the faster it can get merged. -->

#### Describe alternatives you've considered
- Using a database: ew. Enormous dependency, don't need a query language for functionally a key/value store.
- Saving saves as flexbuffer binary: doesn't really get good space savings.

<!-- Explain any alternative solutions, different approaches, or possibilities you've considered using to solve the same problem. -->

#### Testing
Create worlds, run around, save, load, it works. Enable and disable compression on these worlds and others. Intentionally hack the zzip implementation to not save a footer, and confirm the data is successfully recovered through the scan and checksumming. Test compaction works by setting a low compaction threshhold and saving after every step in a game, seeing sizes bounce up and down. Fix many bugs I hit along the way.

![image](https://github.com/user-attachments/assets/30ffbdd6-9b29-4a10-869b-d6a6e4441f57)

Some performance numbers using the quick timer I added in this stack:
Fresh world saving after teleporting around a bunch, writing data both to zzips and to the original files, on a Windows 11 desktop with a Threadripper PRO 7975WX:
```
mapbuffer::save: 3375692us (avg: 3375692us) (count: 1)
  mapbuffer::save_quad: 3151013us (avg: 400us) (count: 7875)
    mapbuffer::save_quad serialize JsonOut: 363504us (avg: 554us) (count: 655)
    mapbuffer::save_quad write to file: 1170192us (avg: 1786us) (count: 655)
    mapbuffer::save_quad write to zzip: 1403013us (avg: 2142us) (count: 655)
      mapbuffer::save_quad compact zzip: 957177us (avg: 1461us) (count: 655)
```

This demonstrates the way the timers can nest. I added a timer at the root of `mapbuffer::save` which encapsulates the whole duration. Breaking it down -
- 0.363s is spent serializing the mapbuffers to strings. This is cost shared in both old and new codepaths.
- 1.17s spent writing the mapbuffers to files in the old codepath
- 1.4s spent writing to zzips in the new codepath, however:
  - 0.975s is spent 'compacting' the zzips to avoid wasting space, but the compaction heuristic is not well tuned right now.
- 0.466s writing zzips exclusive of the compaction, which is more than 2x faster than the old codepath.

After teleporting some more, revisiting previously seen maps, and spending a turn in each location:

```
mapbuffer::save: 5904788us (avg: 2952394us) (count: 2)
  mapbuffer::save_quad: 5469810us (avg: 375us) (count: 14553)
    mapbuffer::save_quad serialize JsonOut: 776581us (avg: 634us) (count: 1223)
    mapbuffer::save_quad write to file: 2203060us (avg: 1801us) (count: 1223)
    mapbuffer::save_quad write to zzip: 2077924us (avg: 1699us) (count: 1223)
      mapbuffer::save_quad compact zzip: 1253596us (avg: 1025us) (count: 1223)
```

For the save above:

```
# Compressed, 'raw' after saving.
Cataclysm-DDA/save/Schaal/maps
$ du -shb .
841843  .

# Uncompressed
Cataclysm-DDA/save/Schaal/maps
$ du -shb .
11948820        .

# Ratio = 11948820 / 841843 = 14.2x

# Recompressed through world option
Cataclysm-DDA/save/Schaal/maps
$ du -shb .
600992  .

# Ratio = 11948820 / 600992 = 19.88x
```

<!-- Describe what steps you took to test that this PR resolved the bug or added the feature, and what tests you performed to make sure it didn't cause any regressions.  Also, include testing suggestions for reviewers and maintainers. See TESTING_YOUR_CHANGES.md -->

#### Additional context

### On the zzip format

`zstd` has a concept of "frames", which are independent chunks of semantically meaningful 'data' to `zstd`. I put data in quotes because there are compressed data frames and also 'skippable' frames. Skippable frames are essentially arbitrary non-compressed data that the zstd library user can encode in a matter that zstd can understand and handle using specific apis.

The `zzip` format encodes some limited per-file metadata using skippable frames, and a richer index as a flexbuffer footer. The flexbuffer format is convenient in that you do not need to know the length of the flexbuffer, you just need to know where it ends. This means we can always put it at the end of the file, and in constant time, read it to access the zzip index. However, for corruption detection, the front of the `zzip` contains a skippable frame containing the length of the footer and its hash. We read this first, then hash the footer, before trusting its contents.

For each file, we encode the filename, a hash of the compressed frame, and then finally the compressed frame itself. We store things in this order to help recover from corruption in case of a crash or power loss event.

| footer length & hash | file name | compressed hash | compressed file | .... | preallocated bytes | flexbuffer footer |
|-|---------|---------|----------------|-|-|-|
|`1234`,`0xABCD` | `cows/steak.json` | `0xDEADBEEF` | `zstd data` | ... | | `index` |

When loading a `zzip`, we try to access the footer. If it is not there, or if the mandatory metadata is missing, we can recover the zzip by scanning it from the front (which is slower than just reading a flexbuffer). In sequence, we can read the filename, the hash of the compressed frame, and then verify the compressed frame is intact using that hash. The zstd frame header contains the frame size, so if we can read the header, we can test the rest of the frame. If at any point there is an issue, we assume the rest of the file is corrupted. Then a fresh footer can be written after the last intact entry.

This assumption is safe because zzips are append-only. To update a file in a zzip, we simply stick the new copy at the end, overwriting the footer if it was in the way, and write a new footer afterward. The old version is orphaned inside the zzip until compaction is triggered. Compaction will write only the latest versions of all files into a new zzip and replace it atomically with filesystem operations, the same way save files were overwritten before.

The compaction algorithm is a little dumb right now. We end up triggering compaction more often than necessary because we first extend the zzip with enough space for the worst case compression result, which ends up padding with a lot of wasted bytes. This causes compaction to happen sooner than the 2x it is currently encoded for. But perfect is the enemy of good, and this is pretty good already right now.

TODO:
- [x] Fix Linux build apparently.
- [x] Debug world / bug report support.
- [x] Map memory.
- [x] Fix reset world removing compression.
- [x] Finish the PR summary.
- [x] Screenshots of compression results.
- [x] Screenshots of performance benchmarks.
- [x] More comments, esp. in code documentation of zzip.
- [x] Overmaps.
- [x] Map memory.
- [x] License notice.
- [x] Android build.
- [x] Linux release build (c/cxx flags mixing improperly in Makefile).
- [x] .sav files.

<!-- Add any other context (such as mock-ups, proof of concepts or screenshots) about the feature or bugfix here. -->


<!--README: Cataclysm: Dark Days Ahead is released under the Creative Commons Attribution ShareAlike 3.0 license.
The code and content of the game are free to use, modify, and redistribute for any purpose whatsoever.
By contributing to the project you agree to the terms of the license and that any contribution you make will also be covered by the same license.
See http://creativecommons.org/licenses/by-sa/3.0/ for details. -->