@ZhilkinSerg 2024-12-30
I've tried zstd with MA overmaps:

| Method | Size  | Disk size |
|--------|--------|--------|
| gz | 8 744 047 | 11 485 184 | 
| zstd -1 | 13 407 208 | 16 068 608 |
| zstd -3 | 12 106 198 | 14 794 752 |
| zstd -3 -D | 12 300 415 | 14 962 688 |
| zstd-19  | 6 893 311  | 9 363 456 |

Dictionary trained with `zstd --train` on all uncompressed omap files with output:

```
Trying 5 different sets of parameters
k=50
d=8
f=20
steps=4
split=75
accel=1
Save dictionary of size 66802 into file dictionary
```

Did not benchmark speed at all.