@akrieger 2024-12-30
Advantage of this format is aggregating sub-4k files together to avoid wasted bytes in the allocation.

I found the best dictionary was with d=20 with the slow cover training params, but on Windows that hits a pessimal behavior in qsort so you should train on windows or fix it like I did (rewrite into c++ and use std::stable_sort parallel).