@akrieger 2024-12-31
> I've tried zstd with MA overmaps:
> ...
> Did not benchmark speed at all.

An interesting thing happens with json around the -13 level (I saw this in development too). With no dictionary:
```
$ zstd/contrib/VS2005/bin/x64/Release/zstd.exe -b1e15 -r MA_overmap/
 1# 1337 files       : 140286061 ->  13401860 (x10.47),  385.8 MB/s,  688.8 MB/s
 2# 1337 files       : 140286061 ->  12869576 (x10.90),  398.4 MB/s,  733.8 MB/s
 3# 1337 files       : 140286061 ->  12100850 (x11.59),  423.8 MB/s,  839.0 MB/s
 4# 1337 files       : 140286061 ->  11375873 (x12.33),  224.8 MB/s,  812.2 MB/s
 5# 1337 files       : 140286061 ->  10260613 (x13.67),  156.1 MB/s, 1218.4 MB/s
 6# 1337 files       : 140286061 ->   9877264 (x14.20),  109.1 MB/s, 1170.0 MB/s
 7# 1337 files       : 140286061 ->   9593546 (x14.62),   68.7 MB/s,  977.8 MB/s
 8# 1337 files       : 140286061 ->   9339853 (x15.02),   54.5 MB/s, 1091.4 MB/s
 9# 1337 files       : 140286061 ->   8858622 (x15.84),   45.3 MB/s, 1195.7 MB/s
10# 1337 files       : 140286061 ->   8444538 (x16.61),   31.6 MB/s, 1214.4 MB/s
11# 1337 files       : 140286061 ->   7770721 (x18.05),   11.0 MB/s, 1820.7 MB/s
12# 1337 files       : 140286061 ->   7568403 (x18.54),   8.83 MB/s, 1816.9 MB/s
13# 1337 files       : 140286061 ->   8586961 (x16.34),   7.04 MB/s, 1342.8 MB/s
14# 1337 files       : 140286061 ->   8044176 (x17.44),   6.06 MB/s, 1972.9 MB/s
15# 1337 files       : 140286061 ->   7033479 (x19.95),   4.57 MB/s, 1962.8 MB/s
```
With dictionary `zstd.exe --train-cover=steps=512,d=20 -r MA_overmap/ -o ma_overmap.dict -M1000MB --maxdict=100KB -T0 -7` (ignore compression/decompression speeds, I changed laptop performance profile in between, just look at ratios):
```
 1# 1337 files       : 140286061 ->  13349825 (x10.51),  664.3 MB/s, 1222.7 MB/s
 2# 1337 files       : 140286061 ->  13690522 (x10.25),  651.0 MB/s, 1210.5 MB/s
 3# 1337 files       : 140286061 ->  12177767 (x11.52),  719.9 MB/s, 1377.6 MB/s
 4# 1337 files       : 140286061 ->  12182817 (x11.52),  618.9 MB/s, 1357.1 MB/s
 5# 1337 files       : 140286061 ->  10714637 (x13.09),  262.3 MB/s  1447.7 MB/s
 6# 1337 files       : 140286061 ->   9801734 (x14.31),  162.1 MB/s, 1901.7 MB/s
 7# 1337 files       : 140286061 ->   9644614 (x14.55),  128.4 MB/s, 1939.1 MB/s
 8# 1337 files       : 140286061 ->   9187718 (x15.27),  106.6 MB/s, 2168.8 MB/s
 9# 1337 files       : 140286061 ->   8702472 (x16.12),   73.8 MB/s, 2345.5 MB/s
10# 1337 files       : 140286061 ->   8274563 (x16.95),   45.5 MB/s, 2595.5 MB/s
11# 1337 files       : 140286061 ->   7556593 (x18.56),   23.8 MB/s, 3356.1 MB/s
12# 1337 files       : 140286061 ->   7322328 (x19.16),   15.9 MB/s, 3476.0 MB/s
13# 1337 files       : 140286061 ->   9136798 (x15.35),   14.8 MB/s, 2419.2 MB/s
14# 1337 files       : 140286061 ->   7879967 (x17.80),   10.8 MB/s, 3081.3 MB/s
15# 1337 files       : 140286061 ->   6781610 (x20.69),   6.44 MB/s, 3444.2 MB/s
```

Thing is, compression speed is irrelevant because the mod content is static. Or static enough. Generally better rations compress better and decompress faster.

Oh and just for fun, as a zzip at compression level 7:
```
$ du -shb save/Timersbaby/maps/ma_overmap.zzip
9776444 save/Timersbaby/maps/ma_overmap.zzip
```

And at 15:
```
$ du -shb save/Timersbaby/maps/ma_overmap.zzip
6913392 save/Timersbaby/maps/ma_overmap.zzip
```