@PatrikLundell 2024-12-31
As far as I understand, this will cause two potential issued:
1. Compressed debug saves. I assume the logic allows for ripping out the unwanted files/folders from a copy of the archive in a modification to the corresponding code (or just copying selected parts), though.
2. Fault finding and hacking. It's not that rare to have the need to examine what's in the save files in order to understand why things misbehave, or remove/fix things that causes a save to blow up when loaded. Thus, there's a need for a tool that allows for the unpacking and repacking of the data in a file. However, having to do that will make the threshold to attempting such activities higher, resulting in poorer support.