@akrieger 2024-12-31
1) yes, good point. However given the size savings it's unclear whether we need to support trimming or just package the dictionary with the subset of bundled maps folders needed. The implementation compresses each subfolder of maps/ separately.
2) there is an option in the main menu for enabling/disabling compression for the whole world at once. It's so fast on normal sized worlds it may as well be instant. For the enormous sample world I have with the 3.5ishGB of data, it takes a couple minutes.