@akrieger 2024-12-31
I wouldn't change anything about the bug report upload, just the maps files inside the save, which would still be a tarball or whatever. Is GitHub complaining about the archive containing those? Like it doesn't support uploading that because it can't virus scan it or something?