@akrieger 2024-12-31
Well yes that is the empirical method, but if I could do that right now I would have :)