@akrieger 2024-12-31
> I'd suggest testing if github is uncooperative by tar:ing up a save compressed with your format and post it in a comment.

[Maunawili.zip](https://github.com/user-attachments/files/18283913/Maunawili.zip)

Yeah github dgaf.