@akrieger 2025-01-17
I split out the zstd source drop into #79220 so this can be more reviewable.