@akrieger 2025-01-18
My thinking is, because the flexbuffer validates the object is an object/dict when we create it, and it does that from the last byte(s) of the file, if that passes then everything before those bytes is also still valid. The initial validation just checks some tag bytes and doesnt seek anywhere else in the file so I believe it should work out.

I'll move the dicts, that makes sense.