@akrieger 2025-01-20
Addressed review comments. Significant refactoring due to inserting a footer checksum frame at the front of the file, DRYing some of frame reading/writing code into private functions, added delete functionality, more comments, just a nontrivial amount got changed.