@akrieger 2025-01-22
Actually, overmaps are easy. They are never deleted and only read/generated in like one place. I can handle those in this PR.

Map memory, maybe slightly more complicated. But I haven't looked too closely. If it's not bad then we can just do everything relevant in this PR and not need any followup.