@sonphantrung 2025-03-06
On a side note, I'm actually highly curious about the origin of the "zzip" format. The closest thing I could find on Google is [this thing](https://github.com/ddebin/zzip/).