@akrieger 2025-03-06
I might rename it to .zstz to avoid confusion but given that one is 9 years old (and I have no plans for making it a separate library) I'm not sure I care.