@akrieger 2025-05-19
I had an idea I want to try for map memory to make it a bit better but otherwise this is pretty good to review.